修复符号错误，打生物崩服的问题

+++ mod/Module.h
@@ -9,6 +9,6 @@ inline std::unordered_map<EventCode, std::vector<PyObject*>> g_callback_function
 //×¢²áÃüÁî
 inline std::unordered_map<std::string, std::pair<std::string, PyObject*>> g_commands;
 //ÉËº¦
-inline static int g_damage = 0;
+inline static float g_damage = 0;
 
 extern "C" PyObject * mc_init();
